QCamera2: Port HAL for 64 bit platforms
- Added -Wextra compiler warning level.
- Avoiding implicit datatype conversions
  and also pointer to integer conversions,
  including literals.
- Avoiding pointer conversions between different types
  other than (void*).
- Using 'size_t' to store memory size and indexes.
- Removed some unused variables.
- Changed datatype of some iterators to match their
  corresponding limiter types.
- Using double instead of float where possible.
- Changed bitmasks to unsigned types.
- Updated printf/scanf format string specifiers to match
  actual type of arguments.
- Using 'String8' instead of 'char[]' where possible
  at QCameraParameters.
- Checking result of QCameraMemory::getSize().
- Fixes build errors and warnings.

CRs-Fixed: 641033

+++ QCamera2/stack/mm-jpeg-interface/src/mm_jpeg_interface.c
@@ -296,7 +296,7 @@ uint32_t jpeg_open(mm_jpeg_ops_t *ops, mm_dimension picture_size)
   char prop[PROPERTY_VALUE_MAX];
 
   property_get("persist.camera.logs", prop, "0");
-  gMmJpegIntfLogLevel = atoi(prop);
+  gMmJpegIntfLogLevel = (uint32_t)atoi(prop);
 
   pthread_mutex_lock(&g_intf_lock);
   /* first time open */
